requirements.txt
PyMuPDF>=1.22.0
python-docx>=0.8.11
Flask==2.3.3
docx2pdf==0.1.8
Flask-Mail==0.9.1
firebase-admin>=6.5.0
python-dotenv>=1.0.1
razorpay>=1.4.2
